# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.PickedOrdersListResponse:
    properties:
      pagination:
        $ref: '#/definitions/utilities.PaginationResponse'
      picked_orders:
        items:
          $ref: '#/definitions/models.PickedOrderResponse'
        type: array
    type: object
...
  models.OrderDetailResponse:
    properties:
      id:
        type: integer
      product:
        allOf:
        - $ref: '#/definitions/models.ProductResponse'
        description: Related data
      product_name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      variant:
        type: string
    type: object
  models.OrderResponse:
    properties:
      address:
        type: string
      buyer:
        type: string
      cancelled_at:
        type: string
      cancelled_by:
        type: string
      changed_at:
        type: string
      changed_by:
        type: string
      channel:
        type: string
      complained:
        type: boolean
      courier:
        type: string
      created_at:
        type: string
      event_status:
        type: string
      id:
        type: integer
      order_details:
        description: Related data
        items:
          $ref: '#/definitions/models.OrderDetailResponse'
        type: array
      order_ginee_id:
        type: string
      pending_at:
        type: string
      pending_by:
        type: string
      picked_at:
        type: string
      picked_by:
        type: string
      processing_status:
        type: string
      sent_before:
        type: string
      store:
        type: string
      tracking:
        type: string
      updated_at:
        type: string
    type: object
...
  models.PickedOrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order:
        allOf:
        - $ref: '#/definitions/models.OrderResponse'
        description: Related data
      order_id:
        type: integer
      picked_by:
        type: integer
      picker:
        $ref: '#/definitions/models.UserResponse'
      updated_at:
        type: string
    type: object
  models.ProductResponse:
    properties:
      barcode:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      location:
        type: string
      name:
        type: string
      sku:
        type: string
      updated_at:
        type: string
      variant:
        type: string
    type: object
...
  models.RoleResponse:
    properties:
      assigned_at:
        type: string
      assigned_by:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      roles:
        items:
          $ref: '#/definitions/models.RoleResponse'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  utilities.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  utilities.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@livotech.com
    name: API Support
  description: A comprehensive user management backend service with JWT authentication
    and role-based access control
  title: Livotech Backend Service API
  version: "1.0"
paths:
...
  /api/picked-orders:
    get:
      consumes:
      - application/json
      description: Get a list of all picked orders with their details and search.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Start date (YYYY-MM-DD format)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD format)
        in: query
        name: end_date
        type: string
      - description: Search by Picker name, Order Ginee ID, or Tracking (partial match)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.PickedOrdersListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utilities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Get all Picked Orders
      tags:
      - Picked-Orders
  /api/picked-orders/{id}:
    get:
      consumes:
      - application/json
      description: Get a picked order by ID.
      parameters:
      - description: Pick order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utilities.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PickedOrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utilities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utilities.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utilities.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utilities.Response'
      security:
      - BearerAuth: []
      summary: Get a picked order by ID
      tags:
      - Picked-Orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"